AffineMotions.jl

Affine motions are functions from a space to a Lie group fulfilling the affine condition.

General Interface

AffineMotions.AffineMotion — Type

struct AffineMotion <: AbstractAffineMotion
    A # action G -> Diff(M)
    f # motion: M -> Alg(G)
    lin # linear part: operator Alg(G) ⊗ Alg(G)*
end

General affine motion, defined from an action of a group $G$ on a manifold $M$, a dynamics $f \colon G \to \mathfrak{g}$, and the corresponding constant linear part lin, a linear endormorphism of $\mathfrak{g}$.

AffineMotions.get_action — Method

get_action(m::AbstractMotion) :: AbstractGroupAction

The action for the given motion.

AffineMotions.get_dynamics — Function

get_dynamics(::AbstractMotion, x) :: Alg(G)

The actual motion x -> φ(x) in Alg(G).

Shortcut: m(x) ≡ get_dynamics(m, x)

AffineMotions.get_lin — Method

get_lin(m::AffineMotion) :: Function

The linear part of the motion, a linear endomorphism of $\mathfrak{G}$.

AffineMotions.get_lin_at — Function

get_lin_at(φ::AbstractMotion, ::Any) :: Function

The linear operator $ξ ↦ ⟨dφ,ξ⟩_x+[φ(x),ξ]$ at the point $x$ for motion $φ$.

AffineMotions.get_lin_mat — Method

Compute matrix for Dφ in the given basis

AffineMotions.integrate — Method

integrate(x0::TM, motion::Motion) :: TM

Integrate the motion motion from the point x0 on the manifold.

AffineMotions.integrate_lift — Method

integrate_lift(x0, m::AbstractMotion) :: AbstractODESolution

Integrate the lifted motion in the acting group.

AffineMotions.morphism — Method

 morphism(φ::Motion, x, B::AbstractBasis)

Return the Lie algebra morphism $\exp(Dφ)$, where $Dφ$ is the linear part of the motion $φ$.

AffineMotions.rescale_motion — Function

rescale_motion(s::Number, φ::Motion) :: Typeof(φ)

Rescale the motion by $s$. This corresponds to the new motion $(sφ)(x) := s(φ(x))$. The motion φ is conveniently rescaled with the syntax s*φ.

AffineMotions.swap_group_motion — Method

swap_group_motion(::Motion)

Swap from standard to dual group action, while preserving the same dynamics.

Base.:* — Method

Compute the tangent vector corresponding to the motion at the point x on the manifold.

Flat Affine Case

AffineMotions.FlatAffineMotion — Type

FlatAffineMotion(A::Matrix,b::Vector)

Creates the affine motion corresponding to the vector field $x ↦ A x + b$.

using AffineMotions
A = [0 1.;0 0]
b = [0, 0]
motion = FlatAffineMotion(A,b)

x0 = [2., 10] # position zero, velocity 10
x1 = integrate(x0, motion)
# output
2-element Vector{Float64}:
 12.0
 10.0

AffineMotions.get_flat_action — Method

get_flat_action(d)

The action associated to a flat space of dimension $d$, namely a translation group acting on a Euclidean space.

Rigid Motions

AffineMotions.RigidMotion — Type

 RigidMotion(action::Action, ξ)

The rigid motion action of a group on a manifold. The variable $ξ$ is an element of the Lie algebra of the acting group base_group(action).

The corresponding vector field $f$ at a point $x$ of the manifold $M$ (i.e., M := group_manifold(action)) is:

\[f(x) = ξ⋅x\]

where $⋅$ represents the infinitesimal action defined by action.

AffineMotions.get_velocity — Function

 get_velocity(m::AbstractRigidMotion)

Access to the underlying infinitesimal motion defining the rigid motion m.

AffineMotions.ZeroMotion — Type

ZeroMotion(::GroupAction)

Trivial rigid motion equal to zero everywhere.

Translations

AffineMotions.TranslationMotion — Method

TranslationMotion(G, ξ, side=LeftSide())

A translation motion $φ : G → Alg(G)$ defined on Lie groups. There is a left and a right version.

Left: $φ(χ) = -χξχ^{-1}$.

Right: $φ(χ) = -χ^{-1}ξχ$.

In both cases, the linear part is the zero operator.

Adjoint Linear Motion

Utilities

AffineMotions.pose_velocities — Method

pose_velocities(G, poses)

Take a list $χ_i$ and returns a list $(χ_i, v_i)$ such that $χ_{i+1} = \exp_{χ_{i}}(v_i)$

AffineMotions.rigid_motion — Method

rigid_motion(G, χ, v, side)

The rigid motion starting at $χ$ in the direction $v$ (tangent to $χ$), using either the standard left action or the dual one.

AffineMotions.AdjointLinearMotion — Type

AdjointLinearMotion(G::DecoratedManifold,M::Array{𝔽, 2},conv::ActionDirection)

One of the affine motions which are neither rigid motions nor translations on the group $SE_k(n)$ (or more general affine groups). If a group element is denoted by $χ = [X;R]$, then the motion corresponding to the $k×k$ matrix $M$ is defined by $φ(χ) := (XM^*;0)$.